create / delete user

+++ modules/delete/story.bash
@@ -1,3 +1,7 @@
-user=$(config user)
-userdel -r -f $user
-id $user 2>/dev/null || echo user removed
+
+user=$(config name)
+
+if test "${user}"; then
+  userdel -r -f $user
+  id $user 2>/dev/null || echo user removed
+fi
